feat: fixed build error
